separat model and controller
